better handling of too-small and too-short models #929
These changes handle small models better in the plater preview.
Extremely small thumbnails are preserved instead of culled,
allowing the model to be scaled up. The scale range is extended to
100000% to allow mm-scale objects expressed in meters to be opened
and scaled to mm. Processing is avoided for models shorter than the
layer height, and more specific warnings are displayed.